+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Added 
 
+
+- Dedicated dataset service for managing datasets  [[GH-79]](https://github.com/delving/hub3/pull/79)
 - Initial service for elasticsearch functionality. Better mapping management.  [[GH-78]](https://github.com/delving/hub3/pull/78)
 - Multitenancy support through organisationIDs  [[GH-77]](https://github.com/delving/hub3/pull/77)
 - Initial version of Time Revision Store  [[GH-76]](https://github.com/delving/hub3/pull/76)